+++ src/main/java/com/oluwaseun/dronedispatch/model/entity/DroneModel.java
@@ -1,4 +1,4 @@
-package com.oluwaseun.dronedispatch.model;
+package com.oluwaseun.dronedispatch.model.entity;
 
 public enum DroneModel {
     LIGHTWEIGHT,
